inference-consumer/requirements.txt
dask[complete]
dask-ml
joblib
kafka-python
numpy
requests
